tests: relocate decoder helpers and unify captured output
Move build_decoder_command into the decoder framework and reuse it
from the encoder, dropping it and the validation flow from base.
Render output via a generic meta["output_sections"] list each
framework populates, so base needs no encoder/decoder knowledge.
The encoder appends its decoder-validation output after its own,
matching runtime order. Extract the encoder validation flow into
_run_decoder_validation_stage.
